tor_2022_group-12/requirements.txt
pandas==1.3.4
nltk==3.6.5
textblob==0.15.3

# Flask Framework
Flask==2.0.2
Flask_Cors==3.0.10

# Gunicorn for daemon
Gunicorn==20.1.0

# Dot env
python_dotenv==0.19.2